requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine learning
scikit-learn>=1.0.0
tensorflow>=2.8.0
keras>=2.8.0
xgboost>=1.5.0
lightgbm>=3.3.0

# EEG signal processing
mne>=1.0.0
pywt>=1.3.0
neurokit2>=0.2.0
antropy>=0.1.4

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
bokeh>=2.4.0

# Data handling
h5py>=3.6.0
joblib>=1.1.0
pickle5>=0.0.11

# Jupyter and development
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0
click>=8.0.0

# Optional: GPU support
# tensorflow-gpu>=2.8.0  # Uncomment if GPU available

# Optional: Advanced signal processing
# librosa>=0.9.0  # For advanced audio-like processing
# pyts>=0.12.0    # For time series feature extraction 